requirements.txt
torch
torchvision
transformers
pandas
numpy
Pillow
pyyaml
tqdm
